Fix unit test
